refactor admin files for docker deployment
