+++ provisioning/Dockerfile
@@ -4,8 +4,10 @@ COPY go.mod go.sum ./
 RUN go mod download
 COPY . .
 RUN CGO_ENABLED=0 GOOS=linux go build -trimpath -o provisioner ./provisioning/cmd/server
+RUN CGO_ENABLED=0 GOOS=linux go build -trimpath -o admin ./provisioning/cmd/admin
 
 FROM scratch
 COPY --from=builder /app/provisioner /provisioner
+COPY --from=builder /app/admin /admin
 EXPOSE 8080
-ENTRYPOINT ["/provisioner"]
+CMD ["/provisioner"]
