+++ provisioning/internal/adminapi/admin_test.go
@@ -1,4 +1,4 @@
-package admin
+package adminapi
 
 import (
 	"bytes"
